+++ src/Waargonaut/Decode/Error.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DeriveFunctor #-}
+-- | Types and typeclass for errors in Waargonaut decoding.
 module Waargonaut.Decode.Error
   ( DecodeError (..)
   , AsDecodeError (..)
@@ -28,7 +29,7 @@ data Err c e
 --
 data DecodeError
   = ConversionFailure Text
-  | KeyDecodeFailed Text
+  | KeyDecodeFailed
   | KeyNotFound Text
   | FailedToMove ZipperMove
   | NumberOutOfBounds JNumber
@@ -36,10 +37,11 @@ data DecodeError
   | ParseFailed Text
   deriving (Show, Eq)
 
+-- | Describes the sorts of errors that may be treated as a 'DecodeError', for use with 'lens'.
 class AsDecodeError r where
   _DecodeError       :: Prism' r DecodeError
   _ConversionFailure :: Prism' r Text
-  _KeyDecodeFailed   :: Prism' r Text
+  _KeyDecodeFailed   :: Prism' r ()
   _KeyNotFound       :: Prism' r Text
   _FailedToMove      :: Prism' r ZipperMove
   _NumberOutOfBounds :: Prism' r JNumber
@@ -65,10 +67,10 @@ instance AsDecodeError DecodeError where
         )
 
   _KeyDecodeFailed
-    = L.prism KeyDecodeFailed
+    = L.prism (const KeyDecodeFailed)
         (\x -> case x of
-            KeyDecodeFailed y -> Right y
-            _                 -> Left x
+            KeyDecodeFailed -> Right ()
+            _               -> Left x
         )
 
   _KeyNotFound