+++ src/Waargonaut/Decode/Internal.hs
@@ -3,9 +3,9 @@
 {-# LANGUAGE FlexibleInstances          #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses      #-}
+{-# LANGUAGE RankNTypes                 #-}
 {-# LANGUAGE TupleSections              #-}
 {-# LANGUAGE TypeFamilies               #-}
-{-# LANGUAGE RankNTypes #-}
 -- | Internal types and functions for building Decoder infrastructure.
 module Waargonaut.Decode.Internal
   ( CursorHistory' (..)
@@ -54,7 +54,7 @@ import           Control.Monad.Except          (ExceptT (..), MonadError (..),
 import           Control.Monad.State           (MonadState (..), StateT (..))
 import           Control.Monad.Trans.Class     (MonadTrans (lift))
 
-import           Control.Monad.Error.Hoist                 ((<!?>))
+import           Control.Monad.Error.Hoist     ((<!?>))
 import           Control.Monad.Morph           (MFunctor (..), MMonad (..))
 
 import           Data.Bifunctor                (first)
@@ -96,6 +96,8 @@ newtype CursorHistory' i = CursorHistory'
   }
   deriving (Show, Eq)
 
+-- |
+-- Pretty print the given 'CursorHistory'' to a more useful format compared to a 'Seq' of 'i'.
 ppCursorHistory
   :: CursorHistory' i
   -> Doc a
@@ -218,6 +220,9 @@ recordZipperMove dir i = L._Wrapped %= (`L.snoc` (dir, i))
 try :: MonadError e m => m a -> m (Maybe a)
 try d = catchError (pure <$> d) (const (pure Nothing))
 
+-- |
+-- Build the basis for a 'Decoder' based on a 'Prism''.
+--
 prismDOrFail'
   :: ( AsDecodeError e
      , MonadError e f