+++ src/Waargonaut/Decode/ZipperMove.hs
@@ -1,3 +1,5 @@
+-- | Types and functions for describing the movements of a cursor around the
+-- 'Waargonaut.Types.Json.Json' structure.
 module Waargonaut.Decode.ZipperMove
   ( ZipperMove (..)
   , AsZipperMove (..)
@@ -10,7 +12,7 @@ import qualified Control.Lens                  as L
 import           Data.Text                     (Text)
 import qualified Data.Text                     as Text
 
-import           Numeric.Natural               (Natural)
+import           Natural                       (Natural)
 
 import           Text.PrettyPrint.Annotated.WL (Doc, (<+>))
 import qualified Text.PrettyPrint.Annotated.WL as WL
@@ -25,22 +27,27 @@ data ZipperMove
   | Item Text
   | L Natural
   | R Natural
+  | BranchFail Text
   deriving (Show, Eq)
 
+-- | Pretty print a given zipper movement, used when printing
+-- 'Waargonaut.Decode.Internal.CursorHistory'' to improve the readability of the errors.
 ppZipperMove :: ZipperMove -> Doc a
 ppZipperMove m = case m of
-  U        -> WL.text "up/"
-  D        -> WL.text "into\\"
+  U              -> WL.text "up/"
+  D              -> WL.text "down\\"
 
-  (L n)    -> WL.text "->-" <+> ntxt n
-  (R n)    -> WL.text "-<-" <+> ntxt n
+  (L n)          -> WL.text "-<-" <+> ntxt n
+  (R n)          -> WL.text "->-" <+> ntxt n
 
-  (DAt k)  -> WL.text "into\\" <+> itxt "key" k
-  (Item t) -> WL.text "-::" <+> itxt "item" t
+  (DAt k)        -> WL.text "into\\" <+> itxt "key" k
+  (Item t)       -> WL.text "-::" <+> itxt "item" t
+  (BranchFail t) -> WL.text "(attempted: " <+> ntxt t <+> WL.text ")"
   where
     itxt t k' = WL.parens (WL.text t <+> WL.colon <+> WL.text (Text.unpack k'))
     ntxt n'   = WL.parens (WL.char 'i' <+> WL.char '+' <+> WL.text (show n'))
 
+-- | Classy 'Control.Lens.Prism''s for things that may be treated as a 'ZipperMove'.
 class AsZipperMove r where
   _ZipperMove :: Prism' r ZipperMove
   _U          :: Prism' r ()