+++ waargbench/src/Waargbench.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE RankNTypes #-}
 module Main (main) where
 
-import           Control.Applicative                          (liftA2, liftA3)
+import           Control.Applicative                          (liftA2)
 import           Control.Lens                                 (isn't, _Left)
 
 import           Data.Functor.Identity                        (Identity)
@@ -18,18 +18,12 @@ import qualified HaskellWorks.Data.Json.Internal.Cursor.Token as HW
 
 import qualified Waargonaut                                   as W
 
-import           Waargonaut.Decode.Error                      (DecodeError)
-
 import           Waargonaut.Decode                            (Decoder)
-import qualified Waargonaut.Decode                            as D
+import qualified Waargonaut.Decode.Traversal                  as D
 
-import qualified Waargonaut.Decode.Succinct                   as SD
+import qualified Waargonaut.Decode                            as SD
 
-import qualified Waargonaut.Generic                           as G
-
-import           Common                                       (Image (..),
-                                                               decodeScientific,
-                                                               imageDecodeGeneric,
+import           Common                                       (decodeScientific, imageDecodeGeneric,
                                                                imageDecodeManual,
                                                                imageDecodeSuccinct,
                                                                parseBS)
@@ -40,20 +34,21 @@ parseOkay = isn't _Left . parseOnly W.parseWaargonaut
 succinctIndexOkay :: ByteString -> Bool
 succinctIndexOkay = isJust . HW.jsonTokenAt . SD.unJCurs . SD.mkCursor
 
-traversalDecode :: Decoder Identity a -> ByteString -> Bool
+traversalDecode :: D.Decoder Identity a -> ByteString -> Bool
 traversalDecode d = isn't _Left . D.simpleDecode parseBS d
 
-succinctDecode :: SD.Decoder Identity a -> ByteString -> Bool
+succinctDecode :: Decoder Identity a -> ByteString -> Bool
 succinctDecode d = isn't _Left . SD.runPureDecode d parseBS . SD.mkCursor
 
 rf :: FilePath -> IO ByteString
 rf f = BS.readFile $ "../test/json-data/" <> f
 
-getParseFiles :: IO (ByteString, ByteString, ByteString)
-getParseFiles = liftA3 (,,)
-  (rf "jp100.json")
-  (rf "twitter100.json")
-  (rf "numbers.json")
+getParseFiles :: IO [ByteString]
+getParseFiles = sequence
+  [ (rf "jp100.json")
+  , (rf "twitter100.json")
+  , (rf "numbers.json")
+  ]
 
 getDecodeFiles :: IO (ByteString, ByteString)
 getDecodeFiles = liftA2 (,)
@@ -78,15 +73,15 @@ decode = G.env getDecodeFiles $ \ ~(image, numbers) -> G.bgroup "Decode"
   ]
 
 parse :: G.Benchmark
-parse = G.env getParseFiles $ \ ~(jp100, twitter100, numbers) -> G.bgroup "Parse - Attoparsec"
-  [ G.bench "jp100"      $ G.nf parseOkay jp100
-  , G.bench "twitter100" $ G.nf parseOkay twitter100
-  , G.bench "numbers"    $ G.nf parseOkay numbers
+parse = G.env getParseFiles $ \ ~(jp100 :twitter100:numbers:_) -> G.bgroup "Parse - Attoparsec"
+  [ G.bench "jp100"                    $ G.nf parseOkay jp100
+  , G.bench "twitter100"               $ G.nf parseOkay twitter100
+  , G.bench "numbers"                  $ G.nf parseOkay numbers
   ]
 
 parseSuccinct :: G.Benchmark
-parseSuccinct = G.env getParseFiles $ \ ~(jp100, twitter100, numbers) -> G.bgroup "Succinct Index"
-  [ G.bench "jp100"      $ G.nf succinctIndexOkay jp100
-  , G.bench "twitter100" $ G.nf succinctIndexOkay twitter100
-  , G.bench "numbers"    $ G.nf succinctIndexOkay numbers
+parseSuccinct = G.env getParseFiles $ \ ~(jp100:twitter100:numbers:_) -> G.bgroup "Succinct Index"
+  [ G.bench "jp100"                    $ G.nf succinctIndexOkay jp100
+  , G.bench "twitter100"               $ G.nf succinctIndexOkay twitter100
+  , G.bench "numbers"                  $ G.nf succinctIndexOkay numbers
   ]